initial fixes
fixed a few bugs, added subtract win/loss buttons and added "back"
buttons to most screens in the admin section.  Probably a bit more work
to do yet but I won't be changing the behavior of the add player screen
like we talked about, it would be way more hassle than it's worth (it
would require making a lot of changes).  Also the "back" button from the
edit error takes you back to player_list, it would take a lot of effort
to get it to return to the edit screen.